Transfer the log content  of create account to HREmployee
