+++ Basecode.Services/Interfaces/IHrEmployeeService.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Basecode.Data.Models;
 using Basecode.Data.Dtos;
+using static Basecode.Services.Services.ErrorHandling;
 
 namespace Basecode.Services.Interfaces
 {
@@ -17,5 +18,6 @@ namespace Basecode.Services.Interfaces
         void Update(HREmployeeUpdationDto hrEmployee);
         void SemiDelete(int id);
         void PermaDelete(int id);
+        public LogContent CreateHrAccount(HREmployeeCreationDto hrEmployee);
     }
 }
