+++ Basecode.Services/Services/HrEmployeeService.cs
@@ -6,10 +6,12 @@ using Basecode.Services.Interfaces;
 
 namespace Basecode.Services.Services
 {
-    public class HrEmployeeService : IHrEmployeeService
+    public class HrEmployeeService : ErrorHandling, IHrEmployeeService
     {
         private readonly IHrEmployeeRepository _repository;
         private readonly IMapper _mapper;
+        private readonly LogContent _logContent = new();
+    
         public HrEmployeeService(IHrEmployeeRepository repository, IMapper mapper) 
         {
             _repository = repository;
@@ -65,13 +67,55 @@ namespace Basecode.Services.Services
             _repository.PermaDelete(id);
         }
 
-
-
         public HrEmployee GetByEmail(string email)
         {
             return _repository.GetByEmail(email);
         }
 
-        
+        public LogContent CreateHrAccount(HREmployeeCreationDto hrEmployee)
+        {
+            if (hrEmployee.Name.Length < 2)
+            {
+                _logContent.Result = false;
+                _logContent.ErrorCode = "ERR! Length Validation";
+                _logContent.Message = "Name must be at least 2 characters long";
+            }
+            else if (hrEmployee.Name.Length < 2)
+            {
+                _logContent.Result = false;
+                _logContent.ErrorCode = "ERR! Length Validation";
+                _logContent.Message = "Name cannot be longer than 150 characters";
+            }
+            else if (!hrEmployee.Email.Contains("@asi-dev2.com"))
+            {
+                _logContent.Result = false;
+                _logContent.ErrorCode = "ERR! Invalid Email";
+                _logContent.Message = "Email is not Alliance Email";
+            }
+            else if (hrEmployee.Email.Length > 50)
+            {
+                _logContent.Result = false;
+                _logContent.ErrorCode = "ERR! Length Validation ";
+                _logContent.Message = "Email cannot be longer than 50 characters long";
+            }
+            else if (hrEmployee.Password!.Length < 6)
+            {
+                _logContent.Result = false;
+                _logContent.ErrorCode = "ERR! Length Validation";
+                _logContent.Message = "Password must be at least 6 characters long";
+            }
+            else if (hrEmployee.Password!.Length > 30)
+            {
+                _logContent.Result = false;
+                _logContent.ErrorCode = "ERR! Length Validation";
+                _logContent.Message = "Password cannot be longer than 30 characters";
+            }
+            else
+            {
+                _logContent.Result = true;
+            }
+
+            return _logContent;
+        }
     }
 }
